Extended Line trace and Hidden Character mesh
Extended the line trace to start in front of the character to prevent the player from shooting behind themselves and created a function that hides the character mesh once the distance between the character and camera becomes too small as the camera transitions itself. Briefly added additive anims for player hit reaction.

Important to note that when the player is pressed against another player the line trace starts behind the enemy player however this only happens when the player is colliding with the other player and at such a close distance it would make more sense to melee as shooting with an assault rifle may prove to be more disorienting for the player at such a close distance.

Keep this in mind when creating the shotgun may need adjusting