Added Collision for SkeletalMesh and Fixed Jitter on Sim Proxies
Created a collision channel for the character mesh so the line trace doesn't detect the capsule component, this allows bullets to only hit the character's body not the air around it blocked by it's capsule component.

Also Smoothed rotation for simulated proxies through the sim proxy rotate function which has it's own properties section. Extracted lines of code that are shared with the AimOffset function into their own respective functions to calculate the player Speed and Aim Offset Pitch